backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1

# Redis and Caching
redis==5.0.1
hiredis==2.2.3

# Background Tasks
celery==5.3.4
flower==2.0.1

# Data Sources
requests==2.31.0
aiohttp==3.9.1

# Historical Data APIs (Phase 2)
yfinance==0.2.28
databento==0.36.0

# Technical Analysis (TA-Lib requires system library - install manually if needed)
# TA-Lib==0.4.28

# Natural Language Processing
textblob==0.17.1
nltk==3.8.1

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2

# Logging and Monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Data Processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Date/Time handling
python-dateutil==2.8.2
pytz==2023.3

# Rate Limiting
slowapi==0.1.9
